back-end/mysite/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
boto3==1.34.127
botocore==1.34.127
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.2
django-admin-interface==0.18.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-decouple==2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.4.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
dnspython==2.5.0
dynaconf==3.2.5
email-validator==2.1.0.post1
et-xmlfile==1.1.0
fastapi==0.109.2
filelock==3.13.1
Flat==0.3.2
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
hvac==2.2.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
jwt==1.3.1
knox==0.1.14
loguru==0.7.2
Markdown==3.5.2
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
passlib==1.7.4
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
s3transfer==0.10.1
setuptools==69.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
sqlmodel==0.0.14
sqlparse==0.4.4
starlette==0.36.3
stripe==9.10.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
validators==0.28.3
virtualenv==20.25.1
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
whitenoise==6.7.0
win32-setctime==1.1.0
zope.event==5.0
zope.interface==6.1
